+++ include/ViZDoomTypes.h
@@ -32,11 +32,11 @@
 
 namespace vizdoom{
 
-    typedef std::vector<uint8_t> Buffer;
-    typedef std::shared_ptr<Buffer> BufferPtr;
+    typedef std::vector<uint8_t> ImageBuffer;
+    typedef std::shared_ptr<ImageBuffer> ImageBufferPtr;
 
-    typedef std::vector<uint16_t> ShortBuffer;
-    typedef std::shared_ptr<ShortBuffer> ShortBufferPtr;
+    typedef std::vector<int16_t> AudioBuffer;
+    typedef std::shared_ptr<AudioBuffer> AudioBufferPtr;
 
     struct Label {
         // Label properties
@@ -107,11 +107,11 @@ namespace vizdoom{
 
         std::vector<double> gameVariables;
 
-        BufferPtr screenBuffer;
-        ShortBufferPtr audioBuffer;
-        BufferPtr depthBuffer;
-        BufferPtr labelsBuffer;
-        BufferPtr automapBuffer;
+        ImageBufferPtr screenBuffer;
+        AudioBufferPtr audioBuffer;
+        ImageBufferPtr depthBuffer;
+        ImageBufferPtr labelsBuffer;
+        ImageBufferPtr automapBuffer;
 
         std::vector<Label> labels;